requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
langchain_huggingface
